portfolio/requirements.txt
asgiref==3.7.2
Django==4.2.5
gunicorn==21.2.0
packaging==23.1
sqlparse==0.4.4
asgiref==3.7.2
Django==4.2.5
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
asgiref==3.7.2
Django==4.2.5
django-crispy-forms==2.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.5
django-crispy-forms==2.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
sqlparse==0.4.4
